Move tests.link.test_lazy tests into tests.test_ifelse
